feat: reorganize the benchmark classes
