+++ benchmarks/src/jmh/kotlin/com/github/kdbc/benchmarks/postgresql/PgBenchmarkBlockingSingleJdbc.kt
@@ -9,10 +9,10 @@ import org.openjdk.jmh.annotations.OutputTimeUnit
 import org.openjdk.jmh.annotations.Scope
 import org.openjdk.jmh.annotations.Setup
 import org.openjdk.jmh.annotations.State
-import org.openjdk.jmh.annotations.TearDown
 import org.openjdk.jmh.annotations.Warmup
 import java.sql.Connection
 import java.util.concurrent.TimeUnit
+import kotlin.use
 
 @Warmup(iterations = 4, time = 10, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 20, time = 10, timeUnit = TimeUnit.SECONDS)
@@ -20,37 +20,49 @@ import java.util.concurrent.TimeUnit
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
-open class JdbcBenchmarkSingle {
+open class PgBenchmarkBlockingSingleJdbc {
     private var id = 0
     private val connection: Connection = getJdbcConnection()
 
     @Setup
     open fun start() {
-        getJdbcConnection().use { connection ->
+        connection.use { connection ->
             connection.createStatement().use { statement ->
                 statement.execute(setupQuery)
             }
         }
     }
 
-    private fun step() {
+    private fun singleStep() {
         id++
         if (id > 5000) id = 1
     }
 
+    private fun multiStep() {
+        id += 10
+        if (id >= 5000) id = 1
+    }
+
     @Benchmark
-    open fun queryData() {
-        step()
-        connection.prepareStatement(jdbcQuerySingle).use { preparedStatement ->
+    open fun queryMultipleRows() {
+        multiStep()
+        connection.prepareStatement(jdbcQuery).use { preparedStatement ->
             preparedStatement.setInt(1, id)
+            preparedStatement.setInt(2, id + 10)
             preparedStatement.executeQuery().use { resultSet ->
                 extractPostDataClassListFromResultSet(resultSet)
             }
         }
     }
 
-    @TearDown
-    open fun destroy() {
-        connection.close()
+    @Benchmark
+    open fun querySingleRow() {
+        singleStep()
+        connection.prepareStatement(jdbcQuerySingle).use { preparedStatement ->
+            preparedStatement.setInt(1, id)
+            preparedStatement.executeQuery().use { resultSet ->
+                extractPostDataClassListFromResultSet(resultSet)
+            }
+        }
     }
 }
