+++ benchmarks/src/jmh/kotlin/com/github/kdbc/benchmarks/postgresql/PgBenchmarkSuspendingSingleKdbc.kt
@@ -1,13 +1,9 @@
 package com.github.kdbc.benchmarks.postgresql
 
-import io.github.clasicrando.kdbc.core.connection.use
+import io.github.clasicrando.kdbc.core.connection.SuspendingConnection
 import io.github.clasicrando.kdbc.core.query.bind
 import io.github.clasicrando.kdbc.core.query.executeClosing
 import io.github.clasicrando.kdbc.core.query.fetchAll
-import io.github.clasicrando.kdbc.postgresql.Postgres
-import io.github.clasicrando.kdbc.postgresql.connection.PgConnectOptions
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.runBlocking
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.BenchmarkMode
@@ -18,6 +14,7 @@ import org.openjdk.jmh.annotations.OutputTimeUnit
 import org.openjdk.jmh.annotations.Scope
 import org.openjdk.jmh.annotations.Setup
 import org.openjdk.jmh.annotations.State
+import org.openjdk.jmh.annotations.TearDown
 import org.openjdk.jmh.annotations.Warmup
 import java.util.concurrent.TimeUnit
 
@@ -27,37 +24,53 @@ import java.util.concurrent.TimeUnit
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
-open class KdbcBenchmarkConcurrentSingle {
+open class PgBenchmarkSuspendingSingleKdbc {
     private var id = 0
-    private val options: PgConnectOptions = runBlocking { initializeConcurrentConnections() }
+    private val connection: SuspendingConnection = runBlocking { getKdbcConnection() }
 
     @Setup
     open fun start(): Unit = runBlocking {
-        Postgres.suspendingConnection(connectOptions = options).use {
-            it.createQuery(setupQuery).executeClosing()
-        }
+        connection.createQuery(setupQuery).executeClosing()
     }
 
-    private fun step(): Int {
+    private fun singleStep(): Int {
         id++
         if (id > 5000) id = 1
         return id
     }
 
-    private suspend fun executeQuery(stepId: Int): List<PostDataClass> {
-        return Postgres.suspendingConnection(connectOptions = options).use { conn ->
-            conn.createPreparedQuery(kdbcQuerySingle)
-                .bind(stepId)
-                .fetchAll(PostDataClassRowParser)
-        }
+    private fun multiStep() {
+        id += 10
+        if (id >= 5000) id = 1
     }
 
     @Benchmark
-    open fun queryData() = runBlocking {
-        val results = List(concurrencyLimit) {
-            val stepId = step()
-            async { executeQuery(stepId) }
+    open fun querySingleRow(): Unit = runBlocking {
+        singleStep()
+        connection.createPreparedQuery(kdbcQuerySingle)
+            .bind(id)
+            .fetchAll(PostDataClassRowParser)
+    }
+
+    @Benchmark
+    open fun queryMultipleRows(): Unit = runBlocking {
+        multiStep()
+        connection.createPreparedQuery(kdbcQuery)
+            .bind(id)
+            .bind(id + 10)
+            .fetchAll(PostDataClassRowParser)
+    }
+
+    @TearDown
+    fun destroy() {
+        runBlocking {
+            if (connection.isConnected) {
+                try {
+                    connection.close()
+                } catch (ex: Throwable) {
+                    ex.printStackTrace()
+                }
+            }
         }
-        results.awaitAll()
     }
 }
